rename the package name of test
