+++ tests/pdctl/region/region_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package region_test
+package region
 
 import (
 	"context"
