+++ tests/pdctl/store/store_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package store_test
+package store
 
 import (
 	"context"
