+++ tests/server/id/id_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package id_test
+package id
 
 import (
 	"context"
