+++ tests/server/join/join_fail/join_fail_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package join_fail_test
+package joinfail
 
 import (
 	"context"
