+++ tests/server/region_syncer/region_syncer_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package syncer_test
+package syncer
 
 import (
 	"context"
