+++ tests/server/tso/allocator_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tso_test
+package tso
 
 import (
 	"context"
