+++ tests/server/watch/leader_watch_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package watch_test
+package watch
 
 import (
 	"context"
